+++ packages/omniston-sdk/src/dto/TransactionRequest.ts
@@ -1,9 +1,13 @@
 import { BuildTransferRequest } from "../api/messages/omni/v1beta7/trader/transaction_builder";
-import type { SetNonNullable } from "../types";
+import type { Converter, SetNonNullable, SetOptional } from "../types";
 
-export type TransactionRequest = SetNonNullable<
-  BuildTransferRequest,
-  "sourceAddress" | "destinationAddress" | "quote"
+export type TransactionRequest = SetOptional<
+  SetNonNullable<
+    BuildTransferRequest,
+    "sourceAddress" | "destinationAddress" | "quote"
+  >,
+  "gasExcessAddress" | "refundAddress"
 >;
 
-export const TransactionRequest = BuildTransferRequest;
+export const TransactionRequest =
+  BuildTransferRequest as Converter<TransactionRequest>;